// lib/api/models/RegisterRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
export type RegisterRequest = {
    /**
     * User email address
     */
    email: string;
    /**
     * User password (minimum 6 characters)
     */
    password: string;
    /**
     * User full name
     */
    name: string;
};